requirements.txt
fastapi-cache2[redis]
streamlit
pandas
plotly
fastapi
requests
uvicorn
snowflake-connector-python
dotenv